requirements.txt
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.28
alpha_vantage>=2.3.1
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0
scipy>=1.11.0
joblib>=1.3.0
requests>=2.31.0
pandas_market_calendars>=4.3.0
pyarrow>=14.0.0
python-dateutil>=2.9.0
pandas-market-calendars>=2.0.0
